+++ src/main/java/me/drex/instantfeedback/datagen/InstantFeedbackDataGenerator.java
@@ -1,6 +1,7 @@
 package me.drex.instantfeedback.datagen;
 
 import me.drex.instantfeedback.entity.ModFrogVariants;
+import me.drex.instantfeedback.entity.ModPigVariants;
 import me.drex.instantfeedback.worldgen.ModVegetationFeatures;
 import me.drex.instantfeedback.worldgen.ModVegetationPlacements;
 import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
@@ -26,5 +27,6 @@ public class InstantFeedbackDataGenerator implements DataGeneratorEntrypoint {
         registryBuilder.add(Registries.CONFIGURED_FEATURE, ModVegetationFeatures::bootstrap);
         registryBuilder.add(Registries.PLACED_FEATURE, ModVegetationPlacements::bootstrap);
         registryBuilder.add(Registries.FROG_VARIANT, ModFrogVariants::bootstrap);
+        registryBuilder.add(Registries.PIG_VARIANT, ModPigVariants::bootstrap);
     }
 }