+++ src/main/java/me/drex/instantfeedback/datagen/ModRegistryProvider.java
@@ -18,6 +18,7 @@ public class ModRegistryProvider extends FabricDynamicRegistryProvider {
         entries.addAll(registries.lookupOrThrow(Registries.PLACED_FEATURE));
         entries.addAll(registries.lookupOrThrow(Registries.CONFIGURED_FEATURE));
         entries.addAll(registries.lookupOrThrow(Registries.FROG_VARIANT));
+        entries.addAll(registries.lookupOrThrow(Registries.PIG_VARIANT));
     }
 
     @Override